server/requirements.txt
transformers
torch
pillow
pdf2image 
torchvision
protobuf
sentencepiece